Change heading-structure of "Sonar project Quality"- and "Sonar Server Vision"-widgets.
Style the elements so that they look like before.
Upgraded gulp to v4.0.2. This is necessary because newer versions of node (currently v14.16.1) do not work with gulp v3.9. It also made it necessary to rewrite the gulpfile, since the execution of tasks changed a lot changed between v3 and v4.